+++ test_common/include/opentelemetry/test_common/ext/http/client/http_client_test_factory.h
@@ -2,7 +2,11 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #pragma once
+
+#include <memory>
+
 #include "opentelemetry/ext/http/client/http_client.h"
+#include "opentelemetry/version.h"
 
 OPENTELEMETRY_BEGIN_NAMESPACE
 namespace ext